rework pension rate representation
Instead of having pension rates be eternal and set in stone, allow them
to change. Update the salary calculate total employee and employer
contributions for the tax year and use the absolute values when
calculating the total costs.

This involves a fairly invasive re-factor of the costs calculator since
we need to cover various complex corner cases like employees changing
salaries in the same years when pension rates change.

Although this code touches a lot of the cost calculation, it doesn't
change any values for simple cases (i.e. where the pension rate doesn't
change) and so one can have some confidence in its correctness since the
tests don't need to be changed.

Since calculate_costs() is a function which is part of the API, we have
to support calling it as it once was so add in an approximation that
contributions are calculated using the rates from the beginning of the
calendar year. This is almost certainly wrong in years when there is a
rate change but is a reasonable approximation for existing code.